feat: Closest body and started work on orbital period calculation
TODO:
Calculate all orbits at the start of the game - save them to a dictionary with a name and a value for the period
Set the UI Text to whichever is selected instead of calculating it every time a planet is clicked.

This will increase initial overhead, but be more performant in the longrun.